+++ tests/testthat/helpers-skip.R
@@ -1,0 +1,5 @@
+skip_if_dtplyr <- function(){
+  if (TRUE) {
+    skip("Not applicable to dtplyr.")
+  }
+}